+++ security/jaas/basic/src/main/resources/WEB-INF/jboss-web.xml
