+++ security/jaas/pom.xml
@@ -15,66 +15,15 @@
     <relativePath>../</relativePath>
   </parent>
 
-  <artifactId>example-jaas</artifactId>
+  <artifactId>example-jaas-parent</artifactId>
 
-  <name>WildFly Swarm Examples: Security JAAS</name>
+  <name>WildFly Swarm Examples: Security JAAS parent</name>
   <description>WildFly Swarm Examples: Security JAAS</description>
 
-  <packaging>jar</packaging>
+  <packaging>pom</packaging>
 
-  <properties>
-    <version.resteasy>3.0.14.Final</version.resteasy>
-  </properties>
-
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.wildfly.swarm</groupId>
-        <artifactId>wildfly-swarm-plugin</artifactId>
-        <configuration>
-          <mainClass>org.wildfly.swarm.examples.jaas.Main</mainClass>
-        </configuration>
-        <executions>
-          <execution>
-            <id>package</id>
-          </execution>
-          <execution>
-            <id>start</id>
-          </execution>
-          <execution>
-            <id>stop</id>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.wildfly.swarm.examples</groupId>
-      <artifactId>examples-base</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.wildfly.swarm</groupId>
-      <artifactId>jaxrs-cdi</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.wildfly.swarm</groupId>
-      <artifactId>jpa</artifactId>
-    </dependency>
-
-    <dependency>
-      <groupId>org.jboss.resteasy</groupId>
-      <artifactId>resteasy-client</artifactId>
-      <version>${version.resteasy}</version>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.jboss.resteasy</groupId>
-      <artifactId>resteasy-jackson2-provider</artifactId>
-      <version>${version.resteasy}</version>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
+  <modules>
+    <module>basic</module>
+  </modules>
 
 </project>
